Moves indenting_ostreambuf out into include/flow
