+++ library/source/flow/indenting_ostreambuf.cpp
@@ -1,4 +1,4 @@
-#include "indenting_ostreambuf.hpp"
+#include "flow/indenting_ostreambuf.hpp"
 
 namespace flow::detail {
